DOC now with elevation and math
